Add log entry verification
Verify that the mapping of package IDs to digests is a 1-1 mapping, and
alert if that mapping becomes inconsistent. This effectively persists a
copy of the log in a map, which gets written to disk after each run of
the monitor as a gob encoding, and decoded on monitor startup.

Also adds regex matching to log matched entries.